requirements.txt
sounddevice
numpy
scipy
openai-whisper
openai
torch
torchvision
torchaudio
PyQt5
python-dotenv